asm: shared: Introduce a common pred_matches_u8 tests
This commit moves pred_matches_u8 tests from x86 as a shared file,
so same test can be also used for aarch64.
